Spring Security, JWT y Autorización.

+++ src/main/java/pe/edu/cibertec/apirestventascibertec/model/response/UsuarioResponse.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 public class UsuarioResponse {
 
-    private Long idusuario;
+    private Integer idusuario;
     private String nomusuario;
     private String token;
